# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-commerce Operations Intelligence Agent API",
    "description": "����������Ӫ Agent ƽ̨ API����ݺ��⻧�������ɷ������֤�߽��ṩ���߷���д��������ͨ������������ƥ�������ա�",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/agent/tasks": {
      "post": {
        "tags": [
          "Agent ����"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_agent_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Agent ����"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_agent_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/tasks/{task_id}": {
      "get": {
        "tags": [
          "Agent ����"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_agent_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "Agent ����"
        ],
        "summary": "Task Events",
        "operationId": "task_events_api_v1_agent_tasks__task_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "Agent ����"
        ],
        "summary": "Cancel Task",
        "operationId": "cancel_task_api_v1_agent_tasks__task_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTask"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentResult": {
        "properties": {
          "result_type": {
            "type": "string",
            "title": "Result Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Facts"
          },
          "inferences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inferences"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          },
          "evidence_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "type": "array",
            "title": "Evidence Refs"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.65
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "result_type",
          "summary"
        ],
        "title": "AgentResult"
      },
      "AgentTask": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "PENDING"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "state_version": {
            "type": "integer",
            "title": "State Version",
            "default": 1
          },
          "request": {
            "$ref": "#/components/schemas/TaskCreate"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "approval_status": {
            "$ref": "#/components/schemas/ApprovalStatus",
            "default": "NOT_REQUIRED"
          },
          "approval_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Hash"
          },
          "approver_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approver Id"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "AgentTask"
      },
      "AgentTaskList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentTask"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AgentTaskList"
      },
      "ApprovalStatus": {
        "type": "string",
        "enum": [
          "NOT_REQUIRED",
          "WAITING_APPROVAL",
          "APPROVED",
          "REJECTED"
        ],
        "title": "ApprovalStatus"
      },
...
      "EvidenceRef": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "grade": {
            "type": "string",
            "pattern": "^[ABCD]$",
            "title": "Grade"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "grade",
          "source",
          "summary"
        ],
        "title": "EvidenceRef"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskCreate": {
        "properties": {
          "tenant_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "user_query": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 5,
            "title": "User Query"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "default": "market_entry"
          },
          "business_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Business Context"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority",
            "default": "MEDIUM"
          }
        },
        "type": "object",
        "required": [
          "user_query"
        ],
        "title": "TaskCreate"
      },
      "TaskPriority": {
        "type": "string",
        "enum": [
          "CRITICAL",
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "TaskPriority"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PLANNING",
          "RUNNING",
          "WAITING_APPROVAL",
          "RETRYING",
          "DEGRADED",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "TaskStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}